Fix offset mouse events in JS/browser
Read window.devicePixelRatio on each event instead of
having its value stored in a lazy that never updates.

This fixes wrongly offset mouse events that could occur when
changing the browser zoom.

We tested that reading window.devicePixelRatio is very cheap.
Executing it is not even measurable in most cases, and
takes 100 microseconds in some cases according to the
measureTime { … } function from the Kotlin stdlib.