Fix scaling issue related to window.devicePixelRatio
Note that this fix will unfortunately not apply to older browser
versions that don't support WeakRef. In such case, a detailed
error will be logged to the console.

Prior to these changes, when window.devicePixelRatio was below 1.0,
which would happen when the webpage zoom was below 50% on high
pixel density screens (aka. HDPi screens), the scaling would be
messed up, leading to wrong coordinates for mouse events.

Also, if you even when window.devicePixelRatio was at 1.0 or above,
zooming the page above 100% would lead to blurry render.

Now, the scale will be updated, and another rendering will be
performed when window.devicePixelRatio changes, which fixes
these issues. Mouse event coordinates will match, and
the visualization will appear as sharp as possible.